Day 14

+++ Node/src/main/java/org/example/service/CreateTable.java
@@ -1,10 +1,11 @@
 package org.example.service;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public interface CreateTable {
     void createTable(String name);
-    void addNoteAboutBuy(String tableName, String key, Integer count, LocalDateTime localDateTime, Float purchase);
+    void addNoteAboutBuy(String tableName, String key, Integer count, LocalDateTime localDateTime, BigDecimal purchase);
     void addNoteAboutSell(String tableName, String key, Integer count);
 
     String getInfoAboutBag(String tableName);
